+++ CHANGELOG.md
@@ -28,6 +28,9 @@ All notable changes to this project will be documented in this file.
   - `DuckDB::Result#use_chunk_each=`
 
 ## Breaking changes
+- The following methods are deprecated.
+  - `DuckDB::Result#_null?`
+  - `DuckDB::Result#row(index)`
 - `DuckDB::Result#use_chunk_each?` is always true.
 - `DuckDB::Result#use_chunk_each=` always ignores the argument.
 - `DuckDB::Result#each` always works the same as `DuckDB::Result#chunk_each`.